refactor: ContentType 페키지 위치 변경
